monitoring: Send config params to skyring for db push
When bigfin attempts to call skyring's module, since
the call is in context of bigfin, the configuration used
is also in the context of bigfin. So, it becomes necessary to
pass the required configuration values from bigfin.
This patch passes the configuration parameters as arguements
to PushToDb methode defined in skyring.

Change-Id: Ib003dd0cd9c898cb4e7f5811ca91eee6aaf859d5
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>