Events: Added handlers for cluster and osd events.
This patch adds handlers for cluster health change events
and osd related events.

Cluster health change:
when skyring gets this event, the status of cluster is updated
accordingly and an alert is raised.

Osd related events:
1. when osd is added or removed: the OSD is added to the DB if
not availabe already, or removed from DB if osd remove event is
received. In both the cases an event(info) is raised.

2. when osd state changes: when osd state changes, the state is
updated accordingly in the DB and an alert is raised.

Change-Id: I31383830d70045549289cdd0921e33c01265f30d
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>