Fix RegExp: Fix the way host was read from DB.
This patch fixes the way in which the host was read from
the DB. Earlier calamari calls used to return trimmed out
host names, so we used to append a "." after the trimmed
out name before searching. But now calamari gives full
hostnames, with this appending "." will create issues, so
removed them in this patch.

Change-Id: I6431c23e784f923fd57a31e4b364c7ff1b618c65
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>