bigfin: Login to calamari if not already logged in
If calamari goes down on a given MON node and we
fall back to other calamari and start, we should
login to the new calamari again.

Introduced a map which maintains logged in status
to calamari of the said MON node and when login
happens its set as true. Rest all mon nodes are
set as false.

Change-Id: Ia01a079f9b784a01f261324fbb529f8ef4f85e97
Signed-off-by: Shubhendu <shtripat@redhat.com>