Bigfin: Change calamari server when server goes down.
This patch starts calamari server on a different mon node when
it receives an event related to calamari server node going down.
Also getCalamariMonNode() method is modified to just return the
calamari mon node, instead of checking the status of calamari
server and handling the failover. Because this is taken care by
event handling and also calamari will automatically start if
the service goes down and node is UP.

Change-Id: I557c1d0ca941cb54f42ab792e0c2d9cb177bf904
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>