Update patch set 1
Patch Set 1: Verified+1

Verified with huge no of OSDs (120) and 13 nodes (3 mons + 10 OSD nodes).
Explicitly I set the calculated value as 20 which is less than 120 (no of OSD), so new value is calculated using the fimula

pgnum = (no of osd / replicas) which is equal to 40. So next 2's power is 64 and the same set as pgnum for the pools.

Patch-set: 1
Label: Verified=+1