RBD-SYNC: Fixed issue in RBD sync when cluster name is not "ceph"
This patch fixes the issue in RBD sync, when the cluster name is not
"ceph"(default cluster name). The reason for this issue was, while
invoking rbd ls command the cluster name was not passed. So if the
cluster name was ceph, the command would succeed, else it would fail.
In this patch the cluster name is passed while invoking the command.

Change-Id: Idc4a506af43bd5e32d541fa9d7e22a278e48b733
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1334470
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>